requirements.txt
Django==3.1.3
djangorestframework==3.12.2
pytest==6.1.2
pytest-django==4.1.0
pytest-cov==2.10.1
python-stdnum==1.14
requests==2.25.0